requirements.txt
Flask==2.2.2
gunicorn==20.1.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.3
google-cloud-logging==3.0.0
